Windows [TP4] Fix docker cp when volumes
